move the video loader to be in the provider so it's not rerenderd if not needed
